requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
packaging==24.2
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.8.2
